+++ R/classSummarisedResult.R
@@ -209,15 +209,12 @@ createSettings <- function(x, settings) {
 
   # remove NA
   colsRemove <- set |>
-    purrr::map(\(x) {
-      if (all(is.na(x))) {
-        x
-      } else {
-        NULL
-      }
-    }) |>
-    purrr::compact() |>
+    purrr::keep(\(x) all(is.na(x))) |>
     names()
+  colsRemove <- colsRemove[!colsRemove %in% c(
+    "result_id", "result_type", "package_name", "package_version", "group",
+    "strata", "additional", "min_cell_count"
+  )]
   if (length(colsRemove) > 0) {
     cli::cli_inform("{.var {colsRemove}} eliminated from settings as all elements are NA.")
     set <- set |>
@@ -944,7 +941,8 @@ transformToSummarisedResult <- function(x,
     tidyr::pivot_longer(
       cols = dplyr::all_of(estimates),
       names_to = "estimate_name",
-      values_to = "estimate_value"
+      values_to = "estimate_value",
+      values_drop_na = TRUE
     ) |>
     dplyr::inner_join(
       types |>